Move internal planarity log to package private api utils, and other minor api utils clean-up.
