+++ c/graphLib/lowLevelUtils/apiutils.h
@@ -13,26 +13,22 @@ extern "C"
 
 #include "stdio.h"
 
-// N.B. Every time this is used to create a string for a message or
-// error message, the developer must check that there will not be a
-// memory overwrite error.
 #define MAXLINE 1024
 
-// N.B. Every time you're trying to read a 32-bit int from a string,
-// you should only need to read this many characters: an optional '-',
-// followed by 10 digits (max signed 32-bit int value is 2,147,483,647).
-// One must always allocate an additional byte for the null-terminator!
+// The string representation for an integer must account for: an optional '-',
+// then 10 digits (max signed 32-bit int), and a null-terminator
 #define MAXCHARSFOR32BITINT 11
 
-    extern int quietMode;
+    // Used within the graphLib and also usable by graph applications to
+    // emit error messages and informational messages.
+    void ErrorMessage(char const *message);
+    void Message(char const *message);
 
-    extern int getQuietModeSetting(void);
-    extern void setQuietModeSetting(int);
+    // These methods control whether ErrorMessage() and Message() calls
+    // emit output or skip producing output
+    int getQuietModeSetting(void);
+    void setQuietModeSetting(int);
 
-    extern void Message(char const *message);
-    extern void ErrorMessage(char const *message);
-
-    int GetNumCharsToReprInt(int theNum, int *numCharsRequired);
 #ifdef __cplusplus
 }
 #endif
