+++ c/planarityApp/planarity.h
@@ -85,6 +85,8 @@ extern "C"
     int ConstructTransformationExpectedResultFilename(char const *infileName, char **outfileName, char command, int actualOrExpectedFlag);
     void WriteAlgorithmResults(graphP theGraph, int Result, char command, platform_time start, platform_time end, char const *infileName);
 
+    int GetNumCharsToReprInt(int theNum, int *numCharsRequired);
+
 #ifdef __cplusplus
 }
 #endif
